Add files via upload
The code calculates and prints the sum of the Taylor Series for a given value of x and a number of terms n provided by user input. The implementation includes the functions fact, powerR, TaylorR, Taylor, and main. The code uses floating-point arithmetic to convert certain values from integer to float.

Below is a brief description of the functionality of the methods:

fact:
The fact function takes an integer value n as an argument, calculates, and returns n!.

powerR:
The powerR function takes a float value x and an integer value n as arguments. It calculates and returns x raised to the power of n, also as a float.

TaylorR:
The TaylorR function takes integer values n and x as arguments. It calculates and returns the sum of the Taylor Series using recursion and the functions fact and powerR. Before calling the powerR function, the program needs to convert the type of the x value to float.

Taylor:
The Taylor function takes integer values n and x as arguments. It calculates and returns the sum of the Taylor Series using iteration and the functions fact and powerR. Before calling the power function, the program needs to convert the type of the x value to float.